+++ SUA/Models/ImpuestosDeduccionesTeatroBorderaux.cs
@@ -8,8 +8,8 @@ namespace SUA.Models
     public class ImpuestosDeduccionesTeatroBorderaux
     {
         public string Nombre { get; set; }
-        public double Porcentaje { get; set; }
-        public double Monto { get; set; }
+        public float Porcentaje { get; set; }
+        public float Monto { get; set; }
         public string Comentarios { get; set; }
     }
 }
